+++ Changes.md
@@ -1,3 +1,4 @@
+* #1230 - Add option to display scoreboard with most recent and all teams scores
 * #1229 - support generating finalist schedule when not running head to head
 * #1222 - Display all performance runs in the performance runs vs. schedule report
 * Fix bug in subjective application where the score summary sometimes would not display